+++ sdk/communication/communication-alpha-ids/recordings/browsers/alphaidsclient__manage_configuration/recording_can_list_all_dynamic_alpha_ids_countries.json
@@ -19,7 +19,7 @@
         "User-Agent": "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) HeadlessChrome/112.0.5614.0 Safari/537.36",
         "x-ms-client-request-id": "00000000-0000-0000-0000-000000000000",
         "x-ms-content-sha256": "47DEQpj8HBSa\u002B/TImW\u002B5JCeuQeRkm5NMpJWZG3hSuFU=",
-        "x-ms-date": "Mon, 26 Jun 2023 18:16:32 GMT",
+        "x-ms-date": "Mon, 26 Jun 2023 21:39:22 GMT",
         "x-ms-useragent": "azsdk-js-communication-alpha-ids/1.0.0-beta.2 core-rest-pipeline/1.11.1 OS"
       },
       "RequestBody": null,
@@ -28,12 +28,12 @@
         "api-supported-versions": "2022-09-26-preview, 2023-07-12",
         "Content-Length": "115",
         "Content-Type": "application/json; charset=utf-8",
-        "Date": "Mon, 26 Jun 2023 18:16:33 GMT",
-        "MS-CV": "7XfiFuek80i2VtQHBaMJSw.0",
+        "Date": "Mon, 26 Jun 2023 21:39:23 GMT",
+        "MS-CV": "E9jdHz\u002BCskC3dz2Q8FwiWA.0",
         "Strict-Transport-Security": "max-age=2592000",
-        "X-Azure-Ref": "0AtaZZAAAAABqs6F0ML5rT4Hu/GAbqTwzREZXMzBFREdFMTYwOQA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
+        "X-Azure-Ref": "0jAWaZAAAAABTsb5nc7NZQ5tyTHO5V6amREZXMzBFREdFMTYxNwA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
         "X-Cache": "CONFIG_NOCACHE",
-        "X-Processing-Time": "254ms"
+        "X-Processing-Time": "342ms"
       },
       "ResponseBody": {
         "countries": [
