+++ sdk/communication/communication-alpha-ids/recordings/browsers/alphaidsclient__manage_configuration/recording_can_manage_configuration.json
@@ -21,7 +21,7 @@
         "User-Agent": "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) HeadlessChrome/112.0.5614.0 Safari/537.36",
         "x-ms-client-request-id": "00000000-0000-0000-0000-000000000000",
         "x-ms-content-sha256": "JrNCayWTdjyW0IkLSnegu/ZtE/xRKwxrE4ojwpDzCio=",
-        "x-ms-date": "Mon, 26 Jun 2023 18:16:28 GMT",
+        "x-ms-date": "Mon, 26 Jun 2023 21:39:20 GMT",
         "x-ms-useragent": "azsdk-js-communication-alpha-ids/1.0.0-beta.2 core-rest-pipeline/1.11.1 OS"
       },
       "RequestBody": {
@@ -32,12 +32,12 @@
         "api-supported-versions": "2022-09-26-preview, 2023-07-12",
         "Content-Length": "16",
         "Content-Type": "application/json; charset=utf-8",
-        "Date": "Mon, 26 Jun 2023 18:16:29 GMT",
-        "MS-CV": "QsRFQTPCYkCnhS0/v5M8UQ.0",
+        "Date": "Mon, 26 Jun 2023 21:39:22 GMT",
+        "MS-CV": "mNRHyK/F8Uq7xsQrhNdnnQ.0",
         "Strict-Transport-Security": "max-age=2592000",
-        "X-Azure-Ref": "0/tWZZAAAAAAPDpv4Wk2KS4N55tr4ha6wREZXMzBFREdFMTYwOQA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
+        "X-Azure-Ref": "0igWaZAAAAADGyurFDBM6RaNLlTYGNT4SREZXMzBFREdFMTYxNwA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
         "X-Cache": "CONFIG_NOCACHE",
-        "X-Processing-Time": "381ms"
+        "X-Processing-Time": "701ms"
       },
       "ResponseBody": {
         "enabled": true
@@ -62,7 +62,7 @@
         "User-Agent": "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) HeadlessChrome/112.0.5614.0 Safari/537.36",
         "x-ms-client-request-id": "00000000-0000-0000-0000-000000000000",
         "x-ms-content-sha256": "47DEQpj8HBSa\u002B/TImW\u002B5JCeuQeRkm5NMpJWZG3hSuFU=",
-        "x-ms-date": "Mon, 26 Jun 2023 18:16:30 GMT",
+        "x-ms-date": "Mon, 26 Jun 2023 21:39:21 GMT",
         "x-ms-useragent": "azsdk-js-communication-alpha-ids/1.0.0-beta.2 core-rest-pipeline/1.11.1 OS"
       },
       "RequestBody": null,
@@ -71,12 +71,12 @@
         "api-supported-versions": "2022-09-26-preview, 2023-07-12",
         "Content-Length": "16",
         "Content-Type": "application/json; charset=utf-8",
-        "Date": "Mon, 26 Jun 2023 18:16:31 GMT",
-        "MS-CV": "akYHbsMayUOyY7rI2MAPtw.0",
+        "Date": "Mon, 26 Jun 2023 21:39:22 GMT",
+        "MS-CV": "LlyVApzSw0WnfhUruVCsOw.0",
         "Strict-Transport-Security": "max-age=2592000",
-        "X-Azure-Ref": "0/9WZZAAAAAAeZzK2iGLET5eLArpSHLRCREZXMzBFREdFMTYwOQA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
+        "X-Azure-Ref": "0iwWaZAAAAAAQdzU8bWT0RYXJziKS0vRkREZXMzBFREdFMTYxNwA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
         "X-Cache": "CONFIG_NOCACHE",
-        "X-Processing-Time": "277ms"
+        "X-Processing-Time": "280ms"
       },
       "ResponseBody": {
         "enabled": true
@@ -103,7 +103,7 @@
         "User-Agent": "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) HeadlessChrome/112.0.5614.0 Safari/537.36",
         "x-ms-client-request-id": "00000000-0000-0000-0000-000000000000",
         "x-ms-content-sha256": "Ws8/93tEIGd7WSMHHzA/rKunqSc6NGKEpmeidd8yUUY=",
-        "x-ms-date": "Mon, 26 Jun 2023 18:16:31 GMT",
+        "x-ms-date": "Mon, 26 Jun 2023 21:39:21 GMT",
         "x-ms-useragent": "azsdk-js-communication-alpha-ids/1.0.0-beta.2 core-rest-pipeline/1.11.1 OS"
       },
       "RequestBody": {
@@ -114,51 +114,12 @@
         "api-supported-versions": "2022-09-26-preview, 2023-07-12",
         "Content-Length": "17",
         "Content-Type": "application/json; charset=utf-8",
-        "Date": "Mon, 26 Jun 2023 18:16:31 GMT",
-        "MS-CV": "aARrXNRlA0WqGPNwSAJFNw.0",
+        "Date": "Mon, 26 Jun 2023 21:39:23 GMT",
+        "MS-CV": "4NPHARdHFUCK1y8uytZZkg.0",
         "Strict-Transport-Security": "max-age=2592000",
-        "X-Azure-Ref": "0ANaZZAAAAACmo4zMSpXKR71DOzQo4R0FREZXMzBFREdFMTYwOQA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
+        "X-Azure-Ref": "0iwWaZAAAAABQuA3yOhocQYqiw1zMgAGXREZXMzBFREdFMTYxNwA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
         "X-Cache": "CONFIG_NOCACHE",
-        "X-Processing-Time": "192ms"
-      },
-      "ResponseBody": {
-        "enabled": false
-      }
-    },
-    {
-      "RequestUri": "https://endpoint/alphaIds/dynamic/configuration?api-version=2023-07-12",
-      "RequestMethod": "GET",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Accept-Encoding": "gzip, deflate, br",
-        "Accept-Language": "en-US",
-        "Authorization": "Sanitized",
-        "Connection": "keep-alive",
-        "Referer": "http://localhost:9876/",
-        "sec-ch-ua": "",
-        "sec-ch-ua-mobile": "?0",
-        "sec-ch-ua-platform": "\u0022\u0022",
-        "Sec-Fetch-Dest": "empty",
-        "Sec-Fetch-Mode": "cors",
-        "Sec-Fetch-Site": "same-site",
-        "User-Agent": "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) HeadlessChrome/112.0.5614.0 Safari/537.36",
-        "x-ms-client-request-id": "00000000-0000-0000-0000-000000000000",
-        "x-ms-content-sha256": "47DEQpj8HBSa\u002B/TImW\u002B5JCeuQeRkm5NMpJWZG3hSuFU=",
-        "x-ms-date": "Mon, 26 Jun 2023 18:16:32 GMT",
-        "x-ms-useragent": "azsdk-js-communication-alpha-ids/1.0.0-beta.2 core-rest-pipeline/1.11.1 OS"
-      },
-      "RequestBody": null,
-      "StatusCode": 200,
-      "ResponseHeaders": {
-        "api-supported-versions": "2022-09-26-preview, 2023-07-12",
-        "Content-Length": "17",
-        "Content-Type": "application/json; charset=utf-8",
-        "Date": "Mon, 26 Jun 2023 18:16:32 GMT",
-        "MS-CV": "nIsVeqaCZ0CFPw3vQNWuQg.0",
-        "Strict-Transport-Security": "max-age=2592000",
-        "X-Azure-Ref": "0AdaZZAAAAAC7wlONbqYDT5C1XSIveQUzREZXMzBFREdFMTYwOQA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
-        "X-Cache": "CONFIG_NOCACHE",
-        "X-Processing-Time": "259ms"
+        "X-Processing-Time": "413ms"
       },
       "ResponseBody": {
         "enabled": false
