+++ sdk/communication/communication-alpha-ids/recordings/browsers/alphaidsclient__preregistered_alpha_ids_operations/recording_can_list_all_preregistered_alpha_ids.json
@@ -19,7 +19,7 @@
         "User-Agent": "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) HeadlessChrome/112.0.5614.0 Safari/537.36",
         "x-ms-client-request-id": "00000000-0000-0000-0000-000000000000",
         "x-ms-content-sha256": "47DEQpj8HBSa\u002B/TImW\u002B5JCeuQeRkm5NMpJWZG3hSuFU=",
-        "x-ms-date": "Mon, 26 Jun 2023 18:16:33 GMT",
+        "x-ms-date": "Mon, 26 Jun 2023 21:39:22 GMT",
         "x-ms-useragent": "azsdk-js-communication-alpha-ids/1.0.0-beta.2 core-rest-pipeline/1.11.1 OS"
       },
       "RequestBody": null,
@@ -28,12 +28,12 @@
         "api-supported-versions": "2022-09-26-preview, 2023-07-12",
         "Content-Length": "15",
         "Content-Type": "application/json; charset=utf-8",
-        "Date": "Mon, 26 Jun 2023 18:16:33 GMT",
-        "MS-CV": "YEp549P88Uiy43\u002BmpFZH5A.0",
+        "Date": "Mon, 26 Jun 2023 21:39:24 GMT",
+        "MS-CV": "\u002BCniR9QVIEGA32UJkR4QfA.0",
         "Strict-Transport-Security": "max-age=2592000",
-        "X-Azure-Ref": "0AtaZZAAAAABPYFbXZb\u002BCQL95rqPXra2RREZXMzBFREdFMTYwOQA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
+        "X-Azure-Ref": "0jAWaZAAAAACMrjUaVQ0UR7TbKlz3XAnnREZXMzBFREdFMTYxNwA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
         "X-Cache": "CONFIG_NOCACHE",
-        "X-Processing-Time": "261ms"
+        "X-Processing-Time": "227ms"
       },
       "ResponseBody": {
         "alphaIds": []
@@ -58,7 +58,7 @@
         "User-Agent": "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) HeadlessChrome/112.0.5614.0 Safari/537.36",
         "x-ms-client-request-id": "00000000-0000-0000-0000-000000000000",
         "x-ms-content-sha256": "47DEQpj8HBSa\u002B/TImW\u002B5JCeuQeRkm5NMpJWZG3hSuFU=",
-        "x-ms-date": "Mon, 26 Jun 2023 18:16:33 GMT",
+        "x-ms-date": "Mon, 26 Jun 2023 21:39:23 GMT",
         "x-ms-useragent": "azsdk-js-communication-alpha-ids/1.0.0-beta.2 core-rest-pipeline/1.11.1 OS"
       },
       "RequestBody": null,
@@ -67,12 +67,12 @@
         "api-supported-versions": "2022-09-26-preview, 2023-07-12",
         "Content-Length": "15",
         "Content-Type": "application/json; charset=utf-8",
-        "Date": "Mon, 26 Jun 2023 18:16:34 GMT",
-        "MS-CV": "FjueEVSDak2efyTxdz3AFQ.0",
+        "Date": "Mon, 26 Jun 2023 21:39:24 GMT",
+        "MS-CV": "DSOum7sjoEew0NjXEUhFpQ.0",
         "Strict-Transport-Security": "max-age=2592000",
-        "X-Azure-Ref": "0AtaZZAAAAAAprDr04UEYT6NuWXntP3\u002BuREZXMzBFREdFMTYwOQA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
+        "X-Azure-Ref": "0jAWaZAAAAABzLsZixf9eSqGZuAVBMFTnREZXMzBFREdFMTYxNwA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
         "X-Cache": "CONFIG_NOCACHE",
-        "X-Processing-Time": "131ms"
+        "X-Processing-Time": "183ms"
       },
       "ResponseBody": {
         "alphaIds": []
