+++ sdk/communication/communication-alpha-ids/recordings/node/alphaidsclient__manage_configuration/recording_can_list_all_dynamic_alpha_ids_countries.json
@@ -11,7 +11,7 @@
         "User-Agent": "azsdk-js-communication-alpha-ids/1.0.0-beta.2 core-rest-pipeline/1.11.1 Node/v18.13.0 OS/(x64-Linux-5.15.90.1-microsoft-standard-WSL2)",
         "x-ms-client-request-id": "00000000-0000-0000-0000-000000000000",
         "x-ms-content-sha256": "47DEQpj8HBSa\u002B/TImW\u002B5JCeuQeRkm5NMpJWZG3hSuFU=",
-        "x-ms-date": "Mon, 26 Jun 2023 18:16:20 GMT"
+        "x-ms-date": "Mon, 26 Jun 2023 21:39:09 GMT"
       },
       "RequestBody": null,
       "StatusCode": 200,
@@ -19,12 +19,12 @@
         "api-supported-versions": "2022-09-26-preview, 2023-07-12",
         "Content-Length": "115",
         "Content-Type": "application/json; charset=utf-8",
-        "Date": "Mon, 26 Jun 2023 18:16:21 GMT",
-        "MS-CV": "N0xTKYmWiU\u002B3o\u002BYLuFlv4A.0",
+        "Date": "Mon, 26 Jun 2023 21:39:11 GMT",
+        "MS-CV": "ZNZw1JlNxkGGdxB51lVRBg.0",
         "Strict-Transport-Security": "max-age=2592000",
-        "X-Azure-Ref": "09dWZZAAAAAAC3BoPsDGKQ6cixRY\u002BVZ9vREZXMzBFREdFMTUwNgA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
+        "X-Azure-Ref": "0fwWaZAAAAADxGLYJq5w0SbxR8gBIhIHtREZXMzBFREdFMTUxNwA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
         "X-Cache": "CONFIG_NOCACHE",
-        "X-Processing-Time": "283ms"
+        "X-Processing-Time": "415ms"
       },
       "ResponseBody": {
         "countries": [
