+++ sdk/communication/communication-alpha-ids/recordings/node/alphaidsclient__manage_configuration/recording_can_manage_configuration.json
@@ -13,7 +13,7 @@
         "User-Agent": "azsdk-js-communication-alpha-ids/1.0.0-beta.2 core-rest-pipeline/1.11.1 Node/v18.13.0 OS/(x64-Linux-5.15.90.1-microsoft-standard-WSL2)",
         "x-ms-client-request-id": "00000000-0000-0000-0000-000000000000",
         "x-ms-content-sha256": "JrNCayWTdjyW0IkLSnegu/ZtE/xRKwxrE4ojwpDzCio=",
-        "x-ms-date": "Mon, 26 Jun 2023 18:16:15 GMT"
+        "x-ms-date": "Mon, 26 Jun 2023 21:39:08 GMT"
       },
       "RequestBody": {
         "enabled": true
@@ -23,12 +23,12 @@
         "api-supported-versions": "2022-09-26-preview, 2023-07-12",
         "Content-Length": "16",
         "Content-Type": "application/json; charset=utf-8",
-        "Date": "Mon, 26 Jun 2023 18:16:17 GMT",
-        "MS-CV": "DieZJJpp1EG/WLhvzmf0hA.0",
+        "Date": "Mon, 26 Jun 2023 21:39:09 GMT",
+        "MS-CV": "zxMpMRb1qUS4EepAgTIkoQ.0",
         "Strict-Transport-Security": "max-age=2592000",
-        "X-Azure-Ref": "08dWZZAAAAABLB903q0/UT6Jwb7qlHqJWREZXMzBFREdFMTUwNgA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
+        "X-Azure-Ref": "0fgWaZAAAAACTcFp7m/\u002BYQYMdvUEYRlFRREZXMzBFREdFMTUxNwA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
         "X-Cache": "CONFIG_NOCACHE",
-        "X-Processing-Time": "768ms"
+        "X-Processing-Time": "422ms"
       },
       "ResponseBody": {
         "enabled": true
@@ -45,7 +45,7 @@
         "User-Agent": "azsdk-js-communication-alpha-ids/1.0.0-beta.2 core-rest-pipeline/1.11.1 Node/v18.13.0 OS/(x64-Linux-5.15.90.1-microsoft-standard-WSL2)",
         "x-ms-client-request-id": "00000000-0000-0000-0000-000000000000",
         "x-ms-content-sha256": "47DEQpj8HBSa\u002B/TImW\u002B5JCeuQeRkm5NMpJWZG3hSuFU=",
-        "x-ms-date": "Mon, 26 Jun 2023 18:16:18 GMT"
+        "x-ms-date": "Mon, 26 Jun 2023 21:39:09 GMT"
       },
       "RequestBody": null,
       "StatusCode": 200,
@@ -53,12 +53,12 @@
         "api-supported-versions": "2022-09-26-preview, 2023-07-12",
         "Content-Length": "16",
         "Content-Type": "application/json; charset=utf-8",
-        "Date": "Mon, 26 Jun 2023 18:16:19 GMT",
-        "MS-CV": "Mp7Zf2L\u002BT0mDhI87Ku20Ow.0",
+        "Date": "Mon, 26 Jun 2023 21:39:10 GMT",
+        "MS-CV": "szIIkg4fUUutMHrPMkOdvg.0",
         "Strict-Transport-Security": "max-age=2592000",
-        "X-Azure-Ref": "089WZZAAAAACv86swWayxSJPsB3WwoY0rREZXMzBFREdFMTUwNgA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
+        "X-Azure-Ref": "0fgWaZAAAAADd\u002Bxpv\u002Bu0VSr/LsBqssGXdREZXMzBFREdFMTUxNwA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
         "X-Cache": "CONFIG_NOCACHE",
-        "X-Processing-Time": "271ms"
+        "X-Processing-Time": "205ms"
       },
       "ResponseBody": {
         "enabled": true
@@ -77,7 +77,7 @@
         "User-Agent": "azsdk-js-communication-alpha-ids/1.0.0-beta.2 core-rest-pipeline/1.11.1 Node/v18.13.0 OS/(x64-Linux-5.15.90.1-microsoft-standard-WSL2)",
         "x-ms-client-request-id": "00000000-0000-0000-0000-000000000000",
         "x-ms-content-sha256": "Ws8/93tEIGd7WSMHHzA/rKunqSc6NGKEpmeidd8yUUY=",
-        "x-ms-date": "Mon, 26 Jun 2023 18:16:18 GMT"
+        "x-ms-date": "Mon, 26 Jun 2023 21:39:09 GMT"
       },
       "RequestBody": {
         "enabled": false
@@ -87,42 +87,12 @@
         "api-supported-versions": "2022-09-26-preview, 2023-07-12",
         "Content-Length": "17",
         "Content-Type": "application/json; charset=utf-8",
-        "Date": "Mon, 26 Jun 2023 18:16:19 GMT",
-        "MS-CV": "aeVJjt1gH06kmxNbhGKnOw.0",
+        "Date": "Mon, 26 Jun 2023 21:39:10 GMT",
+        "MS-CV": "KK18JwBXmEGPw5aCU/fPCw.0",
         "Strict-Transport-Security": "max-age=2592000",
-        "X-Azure-Ref": "089WZZAAAAAAnLGP3cwfbRIwt9moPCZbXREZXMzBFREdFMTUwNgA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
+        "X-Azure-Ref": "0fwWaZAAAAAALySJMTqc5SbniDK2gw6\u002BjREZXMzBFREdFMTUxNwA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
         "X-Cache": "CONFIG_NOCACHE",
-        "X-Processing-Time": "436ms"
-      },
-      "ResponseBody": {
-        "enabled": false
-      }
-    },
-    {
-      "RequestUri": "https://endpoint/alphaIds/dynamic/configuration?api-version=2023-07-12",
-      "RequestMethod": "GET",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Accept-Encoding": "gzip,deflate",
-        "Authorization": "Sanitized",
-        "Connection": "keep-alive",
-        "User-Agent": "azsdk-js-communication-alpha-ids/1.0.0-beta.2 core-rest-pipeline/1.11.1 Node/v18.13.0 OS/(x64-Linux-5.15.90.1-microsoft-standard-WSL2)",
-        "x-ms-client-request-id": "00000000-0000-0000-0000-000000000000",
-        "x-ms-content-sha256": "47DEQpj8HBSa\u002B/TImW\u002B5JCeuQeRkm5NMpJWZG3hSuFU=",
-        "x-ms-date": "Mon, 26 Jun 2023 18:16:20 GMT"
-      },
-      "RequestBody": null,
-      "StatusCode": 200,
-      "ResponseHeaders": {
-        "api-supported-versions": "2022-09-26-preview, 2023-07-12",
-        "Content-Length": "17",
-        "Content-Type": "application/json; charset=utf-8",
-        "Date": "Mon, 26 Jun 2023 18:16:20 GMT",
-        "MS-CV": "U4WrptXN\u002BE2vBX/VQy1lpw.0",
-        "Strict-Transport-Security": "max-age=2592000",
-        "X-Azure-Ref": "09dWZZAAAAACI7XCRnk/PQrNO12v0KlPzREZXMzBFREdFMTUwNgA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
-        "X-Cache": "CONFIG_NOCACHE",
-        "X-Processing-Time": "247ms"
+        "X-Processing-Time": "263ms"
       },
       "ResponseBody": {
         "enabled": false
