+++ sdk/communication/communication-alpha-ids/recordings/node/alphaidsclient__preregistered_alpha_ids_operations/recording_can_list_all_preregistered_alpha_ids.json
@@ -11,7 +11,7 @@
         "User-Agent": "azsdk-js-communication-alpha-ids/1.0.0-beta.2 core-rest-pipeline/1.11.1 Node/v18.13.0 OS/(x64-Linux-5.15.90.1-microsoft-standard-WSL2)",
         "x-ms-client-request-id": "00000000-0000-0000-0000-000000000000",
         "x-ms-content-sha256": "47DEQpj8HBSa\u002B/TImW\u002B5JCeuQeRkm5NMpJWZG3hSuFU=",
-        "x-ms-date": "Mon, 26 Jun 2023 18:16:21 GMT"
+        "x-ms-date": "Mon, 26 Jun 2023 21:39:10 GMT"
       },
       "RequestBody": null,
       "StatusCode": 200,
@@ -19,12 +19,12 @@
         "api-supported-versions": "2022-09-26-preview, 2023-07-12",
         "Content-Length": "15",
         "Content-Type": "application/json; charset=utf-8",
-        "Date": "Mon, 26 Jun 2023 18:16:21 GMT",
-        "MS-CV": "Y/Q4eXQRL0qCbrBOVnB7eQ.0",
+        "Date": "Mon, 26 Jun 2023 21:39:11 GMT",
+        "MS-CV": "hGFvYnCt8Um\u002BJAQA5yaRnA.0",
         "Strict-Transport-Security": "max-age=2592000",
-        "X-Azure-Ref": "09tWZZAAAAAAMzvFCoOvbTaTtEQt\u002BO6wPREZXMzBFREdFMTUwNgA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
+        "X-Azure-Ref": "0gAWaZAAAAADZ7bQlBvurTpVFmsOZwBEAREZXMzBFREdFMTUxNwA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
         "X-Cache": "CONFIG_NOCACHE",
-        "X-Processing-Time": "295ms"
+        "X-Processing-Time": "448ms"
       },
       "ResponseBody": {
         "alphaIds": []
@@ -41,7 +41,7 @@
         "User-Agent": "azsdk-js-communication-alpha-ids/1.0.0-beta.2 core-rest-pipeline/1.11.1 Node/v18.13.0 OS/(x64-Linux-5.15.90.1-microsoft-standard-WSL2)",
         "x-ms-client-request-id": "00000000-0000-0000-0000-000000000000",
         "x-ms-content-sha256": "47DEQpj8HBSa\u002B/TImW\u002B5JCeuQeRkm5NMpJWZG3hSuFU=",
-        "x-ms-date": "Mon, 26 Jun 2023 18:16:21 GMT"
+        "x-ms-date": "Mon, 26 Jun 2023 21:39:10 GMT"
       },
       "RequestBody": null,
       "StatusCode": 200,
@@ -49,12 +49,12 @@
         "api-supported-versions": "2022-09-26-preview, 2023-07-12",
         "Content-Length": "15",
         "Content-Type": "application/json; charset=utf-8",
-        "Date": "Mon, 26 Jun 2023 18:16:22 GMT",
-        "MS-CV": "Dtlf7hJZdkqDEyOlzQG9qw.0",
+        "Date": "Mon, 26 Jun 2023 21:39:12 GMT",
+        "MS-CV": "8sST0o/FgUWZLQbXZLhzGA.0",
         "Strict-Transport-Security": "max-age=2592000",
-        "X-Azure-Ref": "09tWZZAAAAADyeliWAigNQInABr3gqckXREZXMzBFREdFMTUwNgA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
+        "X-Azure-Ref": "0gAWaZAAAAAA5t3\u002BMPkv/QZOk21lWDP11REZXMzBFREdFMTUxNwA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
         "X-Cache": "CONFIG_NOCACHE",
-        "X-Processing-Time": "595ms"
+        "X-Processing-Time": "302ms"
       },
       "ResponseBody": {
         "alphaIds": []
