+++ sdk/communication/communication-alpha-ids/recordings/node/alphaidsclient__preregistered_alpha_ids_operations/recording_can_list_all_preregistered_alpha_ids_countries.json
@@ -11,7 +11,7 @@
         "User-Agent": "azsdk-js-communication-alpha-ids/1.0.0-beta.2 core-rest-pipeline/1.11.1 Node/v18.13.0 OS/(x64-Linux-5.15.90.1-microsoft-standard-WSL2)",
         "x-ms-client-request-id": "00000000-0000-0000-0000-000000000000",
         "x-ms-content-sha256": "47DEQpj8HBSa\u002B/TImW\u002B5JCeuQeRkm5NMpJWZG3hSuFU=",
-        "x-ms-date": "Mon, 26 Jun 2023 18:16:22 GMT"
+        "x-ms-date": "Mon, 26 Jun 2023 21:39:11 GMT"
       },
       "RequestBody": null,
       "StatusCode": 200,
@@ -19,12 +19,12 @@
         "api-supported-versions": "2022-09-26-preview, 2023-07-12",
         "Content-Length": "50",
         "Content-Type": "application/json; charset=utf-8",
-        "Date": "Mon, 26 Jun 2023 18:16:22 GMT",
-        "MS-CV": "/U812CsyR0OqQU7zthPgAw.0",
+        "Date": "Mon, 26 Jun 2023 21:39:12 GMT",
+        "MS-CV": "\u002B/3wMuJsdkm3sZtTdEhm\u002Bg.0",
         "Strict-Transport-Security": "max-age=2592000",
-        "X-Azure-Ref": "099WZZAAAAABu1tByqflKTqN1jgBnnn2FREZXMzBFREdFMTUwNgA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
+        "X-Azure-Ref": "0gQWaZAAAAAA7JnAs2bIRRYk\u002B3ntj\u002BsxRREZXMzBFREdFMTUxNwA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
         "X-Cache": "CONFIG_NOCACHE",
-        "X-Processing-Time": "298ms"
+        "X-Processing-Time": "357ms"
       },
       "ResponseBody": {
         "countries": [
